refactor(ui, apps): update moderation requests
