+++ apps/moderation/src/components/transparency-log.tsx
@@ -3,10 +3,14 @@ import { useTranslation } from 'react-i18next';
 
 import getSDK from '@akashaproject/awf-sdk';
 import DS from '@akashaproject/design-system';
-import { useGetCount, useInfiniteLog } from '@akashaproject/ui-awf-hooks/lib/moderation-request';
+import {
+  ILogItem,
+  useGetCount,
+  useInfiniteLog,
+} from '@akashaproject/ui-awf-hooks/lib/use-moderation';
 
 import Banner from './transparency-log/banner';
-import DetailCard, { ILogItem } from './transparency-log/detail-card';
+import DetailCard from './transparency-log/detail-card';
 
 const { Box, Text, Icon, Spinner, SwitchCard, TransparencyLogMiniCard, useIntersectionObserver } =
   DS;
