+++ apps/moderation/src/components/transparency-log/banner.tsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
+
 import DS from '@akashaproject/design-system';
-import { ICount } from '../content-list';
+import { ICount } from '@akashaproject/ui-awf-hooks/lib/use-moderation';
 
 export interface IBanner {
   count: ICount;
