+++ apps/moderation/src/extensions/moderate-modal.tsx
@@ -5,8 +5,7 @@ import DS from '@akashaproject/design-system';
 import { I18nextProvider, useTranslation } from 'react-i18next';
 import { RootComponentProps } from '@akashaproject/ui-awf-typings';
 import { BrowserRouter as Router } from 'react-router-dom';
-import { useErrors, withProviders } from '@akashaproject/ui-awf-hooks';
-import { useModeration } from '@akashaproject/ui-awf-hooks/lib/moderation-request';
+import { useErrors, useModeration, withProviders } from '@akashaproject/ui-awf-hooks';
 import { BASE_DECISION_URL } from '../services/constants';
 import i18n, { setupI18next } from '../i18n';
 import { useGetLogin } from '@akashaproject/ui-awf-hooks/lib/use-login.new';
