+++ apps/moderation/src/extensions/report-modal.tsx
@@ -10,9 +10,8 @@ import LanguageDetector from 'i18next-browser-languagedetector';
 import { I18nextProvider, initReactI18next, useTranslation } from 'react-i18next';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { RootComponentProps } from '@akashaproject/ui-awf-typings';
-import { useErrors, withProviders, useReasons } from '@akashaproject/ui-awf-hooks';
+import { useErrors, withProviders, useReasons, useModeration } from '@akashaproject/ui-awf-hooks';
 import { BASE_REPORT_URL } from '../services/constants';
-import { useModeration } from '@akashaproject/ui-awf-hooks/lib/moderation-request';
 import { useGetLogin } from '@akashaproject/ui-awf-hooks/lib/use-login.new';
 
 const { ReportModal } = DS;
