+++ ui/hooks/src/constants.ts
@@ -6,10 +6,13 @@ const BASE_DECISION_URL = `${BASE_URL}/decisions`;
 const BASE_MODERATOR_URL = `${BASE_URL}/moderators/status`;
 const BASE_REASONS_URL = `${BASE_URL}/reasons`;
 
+const DEFAULT_FETCH_TIMEOUT = 12000;
+
 export default {
   BASE_REPORT_URL,
   BASE_STATUS_URL,
+  BASE_REASONS_URL,
   BASE_DECISION_URL,
   BASE_MODERATOR_URL,
-  BASE_REASONS_URL,
+  DEFAULT_FETCH_TIMEOUT,
 };
