+++ ui/hooks/src/use-comments.new.ts
@@ -4,7 +4,7 @@ import getSDK from '@akashaproject/awf-sdk';
 import { DataProviderInput } from '@akashaproject/sdk-typings/lib/interfaces/common';
 import { buildPublishObject } from './utils/entry-utils';
 import { logError } from './utils/error-handler';
-import moderationRequest from './moderation-request';
+import { checkStatus } from './use-moderation';
 import { IPublishData, CommentResponse } from '@akashaproject/ui-awf-typings/lib/entry';
 
 export const COMMENT_KEY = 'Comment';
@@ -59,7 +59,7 @@ const getComment = async (commentID): Promise<CommentResponse> => {
   try {
     const user = await lastValueFrom(sdk.api.auth.getCurrentUser());
     // check entry's moderation status
-    const modStatus = await moderationRequest.checkStatus(true, {
+    const modStatus = await checkStatus({
       user: user.data ? user.data.pubKey : '',
       contentIds: [commentID],
     });
